Imported from JSFiddle
This is just the version from the JSFiddle I started working on this in. It
supports basic text fields (Name and Player), Race selection and attributes
with the race selection providing basic modifiers that adjust your base stats.
It doesn't have support for subraces yet, that's next on the list.